requirements.txt
pycryptodome==3.10.1
pandas==1.2.3
flask==1.1.2